Split src/config.rs into src/config/ folder (closes #468)
src/config.rs grew to ~5000 lines covering 22 distinct config structs across
hotkey, audio, output, whisper, text, vad, status, OSD, meeting, notification,
language, profile, and 8 transcription engines. Each new engine repeats the
same four-step recipe and the file's name no longer predicts its contents.
Per the rules in docs/REFACTORING.md this passes both the "third occurrence"
threshold and the "name predicts contents" test.

The split moves each domain into its own file under src/config/. The public
API is unchanged: every type that callers reference (Config, HotkeyConfig,
WhisperConfig, TranscriptionEngine, etc.) is re-exported from
src/config/mod.rs via pub use, so existing `use crate::config::Foo`
statements keep resolving without edit.

Behaviour preservation is checked by counting:

- #[serde( annotations: 179 before / 179 after
- pub struct *Config: 22 / 22
- pub struct (total): 27 / 27
- pub enum (total): 9 / 9
- impl Default for *: 23 / 23
- fn default_* helpers: 51 / 51
- em-dashes introduced: 0

All 770 lib tests pass. cargo fmt and cargo clippy --all-targets --no-deps
-- -D warnings are clean.

Layout:

  src/config/
    mod.rs           Config root, load_config, save_config,
                     parse_config_with_defaults, merge_toml_values,
                     default_true, default_state_file, all pub use
                     re-exports, full tests block (89 #[test] fns)
    default_config.rs DEFAULT_CONFIG + default_config_content
    hotkey.rs        HotkeyConfig, ActivationMode, default_hotkey_key
    audio.rs         AudioConfig, AudioFeedbackConfig
    output.rs        OutputConfig, OutputMode, OutputDriver, FileMode,
                     AppliedLanguageXkbHint, default_language_to_layout
    whisper.rs       WhisperConfig, WhisperMode + whisper helpers
                     (default_on_demand_loading lifted to pub(crate)
                     because the engines reuse it)
    language.rs      LanguageConfig
    text.rs          TextConfig
    vad.rs           VadConfig, VadBackend
    status.rs        StatusConfig + icon themes
    meeting.rs       MeetingConfig + Audio/Diarization/Summary
    notification.rs  NotificationConfig
    profile.rs       Profile + PostProcessConfig
    engines/
      mod.rs         TranscriptionEngine + per-engine pub use
      parakeet.rs    ParakeetConfig + ParakeetModelType
      moonshine.rs
      cohere.rs
      sensevoice.rs
      paraformer.rs
      dolphin.rs
      omnilingual.rs
      soniox.rs
