+++ src/config/audio.rs
@@ -1,0 +1,85 @@
+//! Audio capture and feedback configuration.
+
+use serde::{Deserialize, Serialize};
+
+/// Audio capture configuration
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct AudioConfig {
+    /// PipeWire/PulseAudio device name, or "default"
+    #[serde(default = "default_audio_device")]
+    pub device: String,
+
+    /// Sample rate in Hz (whisper expects 16000)
+    #[serde(default = "default_audio_sample_rate")]
+    pub sample_rate: u32,
+
+    /// Maximum recording duration in seconds (safety limit)
+    #[serde(default = "default_audio_max_duration_secs")]
+    pub max_duration_secs: u32,
+
+    /// Pause MPRIS media players during recording and resume on stop
+    #[serde(default)]
+    pub pause_media: bool,
+
+    /// Audio feedback settings
+    #[serde(default)]
+    pub feedback: AudioFeedbackConfig,
+}
+
+impl Default for AudioConfig {
+    fn default() -> Self {
+        Self {
+            device: default_audio_device(),
+            sample_rate: default_audio_sample_rate(),
+            max_duration_secs: default_audio_max_duration_secs(),
+            pause_media: false,
+            feedback: AudioFeedbackConfig::default(),
+        }
+    }
+}
+
+fn default_audio_device() -> String {
+    "default".to_string()
+}
+
+fn default_audio_sample_rate() -> u32 {
+    16000
+}
+
+fn default_audio_max_duration_secs() -> u32 {
+    60
+}
+
+/// Audio feedback configuration for sound cues
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct AudioFeedbackConfig {
+    /// Enable audio feedback sounds
+    #[serde(default)]
+    pub enabled: bool,
+
+    /// Sound theme: "default", "subtle", "mechanical", or path to custom theme directory
+    #[serde(default = "default_sound_theme")]
+    pub theme: String,
+
+    /// Volume level (0.0 to 1.0)
+    #[serde(default = "default_volume")]
+    pub volume: f32,
+}
+
+fn default_sound_theme() -> String {
+    "default".to_string()
+}
+
+fn default_volume() -> f32 {
+    0.7
+}
+
+impl Default for AudioFeedbackConfig {
+    fn default() -> Self {
+        Self {
+            enabled: false,
+            theme: default_sound_theme(),
+            volume: default_volume(),
+        }
+    }
+}
