+++ src/config/hotkey.rs
@@ -1,0 +1,85 @@
+//! Hotkey detection configuration.
+
+use std::collections::HashMap;
+
+use serde::{Deserialize, Serialize};
+
+use super::default_true;
+
+/// Hotkey activation mode
+#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Default)]
+#[serde(rename_all = "snake_case")]
+pub enum ActivationMode {
+    /// Hold key to record, release to stop (default)
+    #[default]
+    PushToTalk,
+    /// Press once to start recording, press again to stop
+    Toggle,
+}
+
+/// Hotkey detection configuration
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct HotkeyConfig {
+    /// Key name (evdev KEY_* constant name, without the KEY_ prefix)
+    /// Examples: "SCROLLLOCK", "RIGHTALT", "PAUSE", "F24"
+    #[serde(default = "default_hotkey_key")]
+    pub key: String,
+
+    /// Optional modifier keys that must also be held
+    /// Examples: ["LEFTCTRL"], ["LEFTALT", "LEFTSHIFT"]
+    #[serde(default)]
+    pub modifiers: Vec<String>,
+
+    /// Activation mode: push_to_talk (hold to record) or toggle (press to start/stop)
+    #[serde(default)]
+    pub mode: ActivationMode,
+
+    /// Enable built-in hotkey detection (default: true)
+    /// Set to false when using compositor keybindings (Hyprland, Sway) instead
+    /// When disabled, use `voxtype record start/stop/toggle` to control recording
+    #[serde(default = "default_true")]
+    pub enabled: bool,
+
+    /// Optional cancel key (evdev KEY_* constant name, without KEY_ prefix)
+    /// When pressed, cancels the current recording or transcription
+    /// Examples: "ESC", "BACKSPACE", "F12"
+    #[serde(default)]
+    pub cancel_key: Option<String>,
+
+    /// Optional modifier key for secondary model selection (evdev KEY_* name, without KEY_ prefix)
+    /// When held while pressing the hotkey, uses secondary_model instead of the default model
+    /// Examples: "LEFTSHIFT", "RIGHTALT", "LEFTCTRL"
+    #[serde(default)]
+    pub model_modifier: Option<String>,
+
+    /// Optional modifier keys that activate named profiles (evdev KEY_* names, without KEY_ prefix)
+    /// When held while pressing the hotkey, activates the named profile for post-processing
+    /// Example: { "LEFTSHIFT" = "translate" } activates [profiles.translate] when Shift is held
+    #[serde(default)]
+    pub profile_modifiers: HashMap<String, String>,
+}
+
+impl Default for HotkeyConfig {
+    fn default() -> Self {
+        Self {
+            key: default_hotkey_key(),
+            modifiers: Vec::new(),
+            mode: ActivationMode::default(),
+            enabled: true,
+            cancel_key: None,
+            model_modifier: None,
+            profile_modifiers: HashMap::new(),
+        }
+    }
+}
+
+pub(super) fn default_hotkey_key() -> String {
+    #[cfg(target_os = "macos")]
+    {
+        "FN".to_string()
+    }
+    #[cfg(not(target_os = "macos"))]
+    {
+        "SCROLLLOCK".to_string()
+    }
+}
