+++ src/config/mod.rs
@@ -11,355 +11,53 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
-/// Default configuration file content
-pub const DEFAULT_CONFIG: &str = r#"# Voxtype Configuration
-#
-# Location: ~/.config/voxtype/config.toml
-# All settings can be overridden via CLI flags
+mod audio;
+mod default_config;
+pub mod engines;
+mod hotkey;
+mod language;
+mod meeting;
+mod notification;
+mod output;
+mod profile;
+mod status;
+mod text;
+mod vad;
+mod whisper;
 
-# State file for external integrations (Waybar, polybar, etc.)
-# Use "auto" for default location ($XDG_RUNTIME_DIR/voxtype/state),
-# a custom path, or "disabled" to turn off. The daemon writes state
-# ("idle", "recording", "transcribing") to this file whenever it changes.
-# Required for `voxtype record toggle` and `voxtype status` commands.
-state_file = "auto"
+pub use audio::{AudioConfig, AudioFeedbackConfig};
+pub use default_config::{default_config_content, DEFAULT_CONFIG};
+pub use engines::{
+    CohereConfig, DolphinConfig, MoonshineConfig, OmnilingualConfig, ParaformerConfig,
+    ParakeetConfig, ParakeetModelType, SenseVoiceConfig, SonioxConfig, TranscriptionEngine,
+};
+pub use hotkey::{ActivationMode, HotkeyConfig};
+pub use language::LanguageConfig;
+pub use meeting::{
+    MeetingAudioConfig, MeetingConfig, MeetingDiarizationConfig, MeetingSummaryConfig,
+};
+pub use notification::NotificationConfig;
+pub use output::{
+    default_language_to_layout, AppliedLanguageXkbHint, FileMode, OutputConfig, OutputDriver,
+    OutputMode,
+};
+pub use profile::{PostProcessConfig, Profile};
+pub use status::{ResolvedIcons, StatusConfig, StatusIconOverrides};
+pub use text::TextConfig;
+pub use vad::{VadBackend, VadConfig};
+pub use whisper::{WhisperConfig, WhisperMode};
 
-[hotkey]
-# Key to hold for push-to-talk
-# Common choices: SCROLLLOCK, PAUSE, RIGHTALT, F13-F24
-# Use `evtest` to find key names for your keyboard
-key = "SCROLLLOCK"
+#[cfg(test)]
+use hotkey::default_hotkey_key;
+#[cfg(test)]
+use status::load_icon_theme;
 
-# Optional modifier keys that must also be held
-# Example: modifiers = ["LEFTCTRL", "LEFTALT"]
-modifiers = []
-
-# Activation mode: "push_to_talk" or "toggle"
-# - push_to_talk: Hold hotkey to record, release to transcribe (default)
-# - toggle: Press hotkey once to start recording, press again to stop
-# mode = "push_to_talk"
-
-# Enable built-in hotkey detection (default: true)
-# Set to false when using compositor keybindings (Hyprland, Sway) instead
-# When disabled, use `voxtype record start/stop/toggle` to control recording
-# enabled = true
-
-# Modifier key to select secondary model (evdev input mode only)
-# When held while pressing the hotkey, uses whisper.secondary_model instead
-# Example: model_modifier = "LEFTSHIFT"  # Shift+hotkey uses secondary model
-# model_modifier = "LEFTSHIFT"
-
-[audio]
-# Audio input device ("default" uses system default)
-# List devices with: pactl list sources short
-device = "default"
-
-# Sample rate in Hz (whisper expects 16000)
-sample_rate = 16000
-
-# Maximum recording duration in seconds (safety limit)
-max_duration_secs = 60
-
-# Pause MPRIS media players (Spotify, Firefox, etc.) when recording starts,
-# resume them when recording stops. Requires playerctl.
-# pause_media = false
-
-# [audio.feedback]
-# Enable audio feedback sounds (beeps when recording starts/stops)
-# enabled = true
-#
-# Sound theme: "default", "subtle", "mechanical", or path to custom theme directory
-# theme = "default"
-#
-# Volume level (0.0 to 1.0)
-# volume = 0.7
-
-[whisper]
-# Transcription backend: "local" or "remote"
-# - local: Use whisper.cpp locally (default)
-# - remote: Send audio to a remote whisper.cpp server or OpenAI-compatible API
-# backend = "local"
-
-# Model to use for transcription (local backend)
-# Options: tiny, tiny.en, base, base.en, small, small.en, medium, medium.en, large-v3, large-v3-turbo
-# .en models are English-only but faster and more accurate for English
-# large-v3-turbo is faster than large-v3 with minimal accuracy loss (recommended for GPU)
-# Or provide absolute path to a custom .bin model file
-model = "base.en"
-
-# Language for transcription
-# Options:
-#   - Single language: "en", "fr", "de", etc.
-#   - Auto-detect all: "auto"
-#   - Constrained auto-detect: ["en", "fr"] (detects from allowed set only)
-# The array form helps with multilingual users where Whisper might misdetect
-# the language, especially for short sentences.
-# See: https://github.com/openai/whisper#available-models-and-languages
-language = "en"
-
-# Translate non-English speech to English
-translate = false
-
-# Number of CPU threads for inference (omit for auto-detection)
-# threads = 4
-
-# GPU device index for Vulkan/CUDA backend selection.
-# On multi-GPU systems, whisper.cpp may default to the integrated GPU (index 0),
-# causing slower transcription. For different-vendor setups (e.g., Intel iGPU +
-# NVIDIA dGPU), use VOXTYPE_VULKAN_DEVICE=nvidia instead (simpler).
-# Use gpu_device for same-vendor GPUs or precise index control.
-# gpu_device = 1
-
-# Enable flash attention for GPU inference (default: false)
-# Reduces memory usage (~75%) and improves speed (~10%) on CUDA/Vulkan.
-# flash_attention = false
-
-# Initial prompt to provide context for transcription
-# Use this to hint at terminology, proper nouns, or formatting conventions.
-# Example: "Technical discussion about Rust, TypeScript, and Kubernetes."
-# initial_prompt = ""
-
-# --- Multi-model settings ---
-#
-# Secondary model for difficult audio (used with hotkey.model_modifier or CLI --model)
-# secondary_model = "large-v3-turbo"
-#
-# List of available models that can be requested via CLI --model flag
-# available_models = ["large-v3-turbo", "medium.en"]
-#
-# Maximum models to keep loaded in memory (LRU eviction when exceeded)
-# Default: 2 (primary + one secondary). Only applies when gpu_isolation = false.
-# max_loaded_models = 2
-#
-# Seconds before unloading idle secondary models (0 = never auto-unload)
-# Default: 300 (5 minutes). Only applies when gpu_isolation = false.
-# cold_model_timeout_secs = 300
-
-# --- Eager processing settings ---
-#
-# Enable eager input processing (transcribe chunks while recording continues)
-# Reduces perceived latency on slower machines by processing audio in parallel.
-# eager_processing = false
-#
-# Duration of each audio chunk in seconds (default: 5.0)
-# eager_chunk_secs = 5.0
-#
-# Overlap between chunks in seconds (helps catch words at boundaries, default: 0.5)
-# eager_overlap_secs = 0.5
-
-# --- Remote backend settings (used when backend = "remote") ---
-#
-# Remote server endpoint URL (required for remote backend)
-# Examples:
-#   - whisper.cpp server: "http://192.168.1.100:8080"
-#   - OpenAI API: "https://api.openai.com"
-# remote_endpoint = "http://192.168.1.100:8080"
-#
-# Model name to send to remote server (default: "whisper-1")
-# remote_model = "whisper-1"
-#
-# API key for remote server (optional, or use VOXTYPE_WHISPER_API_KEY env var)
-# remote_api_key = ""
-#
-# Timeout for remote requests in seconds (default: 30)
-# remote_timeout_secs = 30
-
-[output]
-# Primary output mode: "type" or "clipboard"
-# - type: Simulates keyboard input at cursor position (requires ydotool)
-# - clipboard: Copies text to clipboard (requires wl-copy)
-mode = "type"
-
-# Fall back to clipboard if typing fails
-fallback_to_clipboard = true
-
-# Custom driver order for type mode (optional)
-# Default order: wtype -> dotool -> ydotool -> clipboard
-# Customize to prefer a specific driver or change the fallback order.
-# Available drivers: wtype, dotool, ydotool, clipboard
-# Example: prefer ydotool over dotool:
-#   driver_order = ["wtype", "ydotool", "dotool", "clipboard"]
-# Example: use only ydotool, no fallback:
-#   driver_order = ["ydotool"]
-# driver_order = ["wtype", "dotool", "ydotool", "clipboard"]
-
-# Per-language XKB layout variants for multilingual dictation.
-# Use this with language arrays such as `language = ["en", "ru"]` when a
-# language needs a variant that should not apply to other languages.
-# Built-in layout defaults already include `ru = "ru"` in language_to_layout.
-# [output.language_to_variant]
-# ru = "phonetic"
-
-# Delay between typed characters in milliseconds
-# 0 = fastest possible, increase if characters are dropped
-type_delay_ms = 0
-
-# Automatically submit (send Enter key) after outputting transcribed text
-# Useful for chat applications, command lines, or forms where you want
-# to auto-submit after dictation
-# auto_submit = true
-
-# Convert newlines to Shift+Enter instead of regular Enter
-# Useful for applications where Enter submits (e.g., Cursor IDE, Slack, Discord)
-# shift_enter_newlines = false
-
-# Prefix wtype output with a Shift key press/release
-# Workaround for apps (e.g., Discord) that drop the first CJK character
-# wtype_shift_prefix = false
-
-# Restore clipboard content after paste mode (default: false)
-# Saves clipboard before transcription, restores it after paste keystroke
-# Only applies to mode = "paste". Useful when you want to preserve your
-# existing clipboard content across dictation operations.
-# restore_clipboard = false
-
-# Delay after paste before restoring clipboard (milliseconds)
-# Allows time for the paste operation to complete (default: 200)
-# restore_clipboard_delay_ms = 200
-
-# Wait for modifier keys to be released before typing (default: true)
-# Prevents transcribed letters from combining with held modifiers (Ctrl/Alt/
-# Shift/Super) and triggering compositor or application keybindings. Reads
-# modifier state via evdev; requires user to be in the 'input' group.
-# Silently disabled when /dev/input is unreadable, so output proceeds as
-# before in that case.
-# wait_for_modifier_release = true
-
-# Maximum milliseconds to wait for modifier release before falling back to
-# clipboard output. Prevents a stuck modifier from indefinitely blocking
-# transcription delivery. (default: 750)
-# modifier_release_timeout_ms = 750
-
-# Pre/post output hooks (optional)
-# Commands to run before and after typing output. Useful for compositor integration.
-# Example: Block modifier keys during typing with Hyprland submap:
-#   pre_output_command = "hyprctl dispatch submap voxtype_suppress"
-#   post_output_command = "hyprctl dispatch submap reset"
-# See troubleshooting docs for the required Hyprland submap configuration.
-# Note: usually unnecessary now that wait_for_modifier_release is enabled by
-# default; the submap workaround is only needed if /dev/input is unreadable.
-
-# Post-processing command (optional)
-# Pipe transcribed text through an external command for cleanup before output.
-# The command receives text on stdin and outputs processed text on stdout.
-# Useful for LLM-based text cleanup, grammar correction, filler word removal.
-# On any failure (timeout, error), falls back to original transcription.
-#
-# [output.post_process]
-# command = "ollama run llama3.2:1b 'Clean up this dictation. Fix grammar, remove filler words. Output only the cleaned text:'"
-# timeout_ms = 30000  # 30 second timeout (generous for LLM)
-# trim = true         # Strip leading/trailing whitespace from output (default: true)
-# fallback_on_empty = true  # Use original text if command returns empty (default: true)
-
-[output.notification]
-# Show notification when recording starts (hotkey pressed)
-on_recording_start = false
-
-# Show notification when recording stops (transcription beginning)
-on_recording_stop = false
-
-# Show notification with transcribed text after transcription completes
-on_transcription = true
-
-# Notification urgency level: "low", "normal", or "critical".
-# On GNOME, "low" notifications are delivered to the drawer without a popup banner.
-# Use "normal" (default) to ensure notifications appear as banners.
-# urgency = "normal"
-
-# [text]
-# Text processing options (word replacements, spoken punctuation)
-#
-# Enable spoken punctuation conversion (e.g., say "period" to get ".")
-# spoken_punctuation = false
-#
-# Custom word replacements (case-insensitive)
-# replacements = { "vox type" = "voxtype" }
-#
-# Smart auto-submit: say "submit" at the end of dictation to press Enter.
-# The word "submit" is stripped from the output text and Enter is pressed.
-# smart_auto_submit = false
-#
-# Remove filler words like "uh" and "um" from transcribed text.
-# Enabled by default. Set filter_filler_words = false to disable, or override
-# the word list via filler_words.
-# filter_filler_words = true
-# filler_words = ["uh", "um", "er", "ah", "eh", "hmm", "hm", "mm", "mhm"]
-
-# [vad]
-# Voice Activity Detection - filters silence-only recordings
-# Prevents Whisper hallucinations on silent audio
-#
-# enabled = false      # Enable VAD (off by default)
-# threshold = 0.5      # 0.0 = sensitive, 1.0 = aggressive
-# min_speech_duration_ms = 100  # Minimum speech required
-
-# [status]
-# Status display icons for Waybar/tray integrations
-#
-# Icon theme (or path to custom theme file):
-#   Font-based (require specific fonts):
-#     - "emoji"     - Default emoji icons (🎙️ 🎤 ⏳)
-#     - "nerd-font" - Nerd Font icons (requires Nerd Font)
-#     - "material"  - Material Design Icons (requires MDI font)
-#     - "phosphor"  - Phosphor Icons (requires Phosphor font)
-#     - "codicons"  - VS Code icons (requires Codicons font)
-#     - "omarchy"   - Omarchy distro icons
-#   Universal (no special fonts needed):
-#     - "minimal"   - Simple Unicode (○ ● ◐ ×)
-#     - "dots"      - Geometric shapes (◯ ⬤ ◔ ◌)
-#     - "arrows"    - Media player style (▶ ● ↻ ■)
-#     - "text"      - Plain text ([MIC] [REC] [...] [OFF])
-# icon_theme = "emoji"
-#
-# Per-state icon overrides (optional, takes precedence over theme)
-# [status.icons]
-# idle = "🎙️"
-# recording = "🎤"
-# transcribing = "⏳"
-# stopped = ""
-
-# [profiles]
-# Named profiles for context-specific post-processing
-# Use with: voxtype record start --profile slack
-#
-# [profiles.slack]
-# post_process_command = "ollama run llama3.2:1b 'Format for Slack...'"
-#
-# [profiles.code]
-# post_process_command = "ollama run llama3.2:1b 'Format as code comment...'"
-# output_mode = "clipboard"
-"#;
-
-/// Return the default config content with platform-appropriate hotkey
-pub fn default_config_content() -> String {
-    #[cfg(target_os = "macos")]
-    {
-        DEFAULT_CONFIG
-            .replace("key = \"SCROLLLOCK\"", "key = \"FN\"")
-            .replace(
-                "# Common choices: SCROLLLOCK, PAUSE, RIGHTALT, F13-F24",
-                "# Common choices: FN, RIGHTALT, F13-F24",
-            )
-            .replace(
-                "# Use `evtest` to find key names for your keyboard",
-                "# FN (Globe key) is recommended on macOS",
-            )
-    }
-    #[cfg(not(target_os = "macos"))]
-    {
-        DEFAULT_CONFIG.to_string()
-    }
+pub(super) fn default_true() -> bool {
+    true
 }
 
-/// Hotkey activation mode
-#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum ActivationMode {
-    /// Hold key to record, release to stop (default)
-    #[default]
-    PushToTalk,
-    /// Press once to start recording, press again to stop
-    Toggle,
+fn default_state_file() -> Option<String> {
+    Some("auto".to_string())
 }
 
 /// Root configuration structure
@@ -446,1997 +144,6 @@ pub struct Config {
     /// Use with: `voxtype record start --profile slack`
     #[serde(default)]
     pub profiles: HashMap<String, Profile>,
-}
-
-/// Hotkey detection configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct HotkeyConfig {
-    /// Key name (evdev KEY_* constant name, without the KEY_ prefix)
-    /// Examples: "SCROLLLOCK", "RIGHTALT", "PAUSE", "F24"
-    #[serde(default = "default_hotkey_key")]
-    pub key: String,
-
-    /// Optional modifier keys that must also be held
-    /// Examples: ["LEFTCTRL"], ["LEFTALT", "LEFTSHIFT"]
-    #[serde(default)]
-    pub modifiers: Vec<String>,
-
-    /// Activation mode: push_to_talk (hold to record) or toggle (press to start/stop)
-    #[serde(default)]
-    pub mode: ActivationMode,
-
-    /// Enable built-in hotkey detection (default: true)
-    /// Set to false when using compositor keybindings (Hyprland, Sway) instead
-    /// When disabled, use `voxtype record start/stop/toggle` to control recording
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-
-    /// Optional cancel key (evdev KEY_* constant name, without KEY_ prefix)
-    /// When pressed, cancels the current recording or transcription
-    /// Examples: "ESC", "BACKSPACE", "F12"
-    #[serde(default)]
-    pub cancel_key: Option<String>,
-
-    /// Optional modifier key for secondary model selection (evdev KEY_* name, without KEY_ prefix)
-    /// When held while pressing the hotkey, uses secondary_model instead of the default model
-    /// Examples: "LEFTSHIFT", "RIGHTALT", "LEFTCTRL"
-    #[serde(default)]
-    pub model_modifier: Option<String>,
-
-    /// Optional modifier keys that activate named profiles (evdev KEY_* names, without KEY_ prefix)
-    /// When held while pressing the hotkey, activates the named profile for post-processing
-    /// Example: { "LEFTSHIFT" = "translate" } activates [profiles.translate] when Shift is held
-    #[serde(default)]
-    pub profile_modifiers: HashMap<String, String>,
-}
-
-impl Default for HotkeyConfig {
-    fn default() -> Self {
-        Self {
-            key: default_hotkey_key(),
-            modifiers: Vec::new(),
-            mode: ActivationMode::default(),
-            enabled: true,
-            cancel_key: None,
-            model_modifier: None,
-            profile_modifiers: HashMap::new(),
-        }
-    }
-}
-
-/// Audio capture configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct AudioConfig {
-    /// PipeWire/PulseAudio device name, or "default"
-    #[serde(default = "default_audio_device")]
-    pub device: String,
-
-    /// Sample rate in Hz (whisper expects 16000)
-    #[serde(default = "default_audio_sample_rate")]
-    pub sample_rate: u32,
-
-    /// Maximum recording duration in seconds (safety limit)
-    #[serde(default = "default_audio_max_duration_secs")]
-    pub max_duration_secs: u32,
-
-    /// Pause MPRIS media players during recording and resume on stop
-    #[serde(default)]
-    pub pause_media: bool,
-
-    /// Audio feedback settings
-    #[serde(default)]
-    pub feedback: AudioFeedbackConfig,
-}
-
-impl Default for AudioConfig {
-    fn default() -> Self {
-        Self {
-            device: default_audio_device(),
-            sample_rate: default_audio_sample_rate(),
-            max_duration_secs: default_audio_max_duration_secs(),
-            pause_media: false,
-            feedback: AudioFeedbackConfig::default(),
-        }
-    }
-}
-
-fn default_audio_device() -> String {
-    "default".to_string()
-}
-
-fn default_audio_sample_rate() -> u32 {
-    16000
-}
-
-fn default_audio_max_duration_secs() -> u32 {
-    60
-}
-
-/// Audio feedback configuration for sound cues
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct AudioFeedbackConfig {
-    /// Enable audio feedback sounds
-    #[serde(default)]
-    pub enabled: bool,
-
-    /// Sound theme: "default", "subtle", "mechanical", or path to custom theme directory
-    #[serde(default = "default_sound_theme")]
-    pub theme: String,
-
-    /// Volume level (0.0 to 1.0)
-    #[serde(default = "default_volume")]
-    pub volume: f32,
-}
-
-fn default_hotkey_key() -> String {
-    #[cfg(target_os = "macos")]
-    {
-        "FN".to_string()
-    }
-    #[cfg(not(target_os = "macos"))]
-    {
-        "SCROLLLOCK".to_string()
-    }
-}
-
-fn default_sound_theme() -> String {
-    "default".to_string()
-}
-
-fn default_volume() -> f32 {
-    0.7
-}
-
-fn default_on_demand_loading() -> bool {
-    false
-}
-
-fn default_context_window_optimization() -> bool {
-    false
-}
-
-fn default_max_loaded_models() -> usize {
-    2 // Primary model + one secondary
-}
-
-fn default_cold_model_timeout() -> u64 {
-    300 // 5 minutes
-}
-
-fn default_eager_chunk_secs() -> f32 {
-    5.0
-}
-
-fn default_eager_overlap_secs() -> f32 {
-    0.5
-}
-
-fn default_whisper_model() -> String {
-    "base.en".to_string()
-}
-
-fn default_state_file() -> Option<String> {
-    Some("auto".to_string())
-}
-
-impl Default for AudioFeedbackConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            theme: default_sound_theme(),
-            volume: default_volume(),
-        }
-    }
-}
-
-/// Status display configuration for Waybar/tray integrations
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct StatusConfig {
-    /// Icon theme: "emoji", "nerd-font", "omarchy", "minimal", or path to custom theme
-    #[serde(default = "default_icon_theme")]
-    pub icon_theme: String,
-
-    /// Per-state icon overrides (optional, takes precedence over theme)
-    #[serde(default)]
-    pub icons: StatusIconOverrides,
-}
-
-fn default_icon_theme() -> String {
-    "emoji".to_string()
-}
-
-impl Default for StatusConfig {
-    fn default() -> Self {
-        Self {
-            icon_theme: default_icon_theme(),
-            icons: StatusIconOverrides::default(),
-        }
-    }
-}
-
-/// Per-state icon overrides for status display
-#[derive(Debug, Clone, Default, Deserialize, Serialize)]
-pub struct StatusIconOverrides {
-    pub idle: Option<String>,
-    pub recording: Option<String>,
-    pub streaming: Option<String>,
-    pub transcribing: Option<String>,
-    pub stopped: Option<String>,
-}
-
-/// Resolved icons for each state (after applying theme + overrides)
-#[derive(Debug, Clone)]
-pub struct ResolvedIcons {
-    pub idle: String,
-    pub recording: String,
-    pub streaming: String,
-    pub transcribing: String,
-    pub stopped: String,
-}
-
-impl StatusConfig {
-    /// Resolve icons by loading theme and applying any overrides
-    pub fn resolve_icons(&self) -> ResolvedIcons {
-        // Start with theme defaults
-        let mut icons = load_icon_theme(&self.icon_theme);
-
-        // Apply per-state overrides
-        if let Some(ref icon) = self.icons.idle {
-            icons.idle = icon.clone();
-        }
-        if let Some(ref icon) = self.icons.recording {
-            icons.recording = icon.clone();
-        }
-        if let Some(ref icon) = self.icons.streaming {
-            icons.streaming = icon.clone();
-        }
-        if let Some(ref icon) = self.icons.transcribing {
-            icons.transcribing = icon.clone();
-        }
-        if let Some(ref icon) = self.icons.stopped {
-            icons.stopped = icon.clone();
-        }
-
-        icons
-    }
-}
-
-/// Load an icon theme by name or from a custom file path
-fn load_icon_theme(theme: &str) -> ResolvedIcons {
-    match theme {
-        "emoji" => ResolvedIcons {
-            idle: "🎙️".to_string(),
-            recording: "🎤".to_string(),
-            streaming: "📡".to_string(), // satellite antenna — live broadcast
-            transcribing: "⏳".to_string(),
-            stopped: "".to_string(),
-        },
-        "nerd-font" => ResolvedIcons {
-            // Nerd Font icons: microphone, circle, spinner, microphone-slash
-            idle: "\u{f130}".to_string(),         // nf-fa-microphone
-            recording: "\u{f111}".to_string(),    // nf-fa-circle (filled)
-            streaming: "\u{f519}".to_string(),    // nf-fa-broadcast_tower
-            transcribing: "\u{f110}".to_string(), // nf-fa-spinner
-            stopped: "\u{f131}".to_string(),      // nf-fa-microphone_slash
-        },
-        "omarchy" => ResolvedIcons {
-            // Material Design icons matching Omarchy waybar config
-            idle: "\u{ec12}".to_string(), // nf-md-microphone_outline
-            recording: "\u{f036c}".to_string(), // nf-md-microphone
-            streaming: "\u{f048b}".to_string(), // nf-md-access_point — broadcasting/live
-            transcribing: "\u{f051f}".to_string(), // nf-md-timer_sand
-            stopped: "\u{ec12}".to_string(), // nf-md-microphone_outline
-        },
-        "minimal" => ResolvedIcons {
-            idle: "○".to_string(),
-            recording: "●".to_string(),
-            streaming: "⊙".to_string(), // U+2299 circled dot — active/live
-            transcribing: "◐".to_string(),
-            stopped: "×".to_string(),
-        },
-        "material" => ResolvedIcons {
-            // Material Design Icons (requires MDI font)
-            idle: "\u{f036c}".to_string(),         // mdi-microphone
-            recording: "\u{f040a}".to_string(),    // mdi-record
-            streaming: "\u{f048b}".to_string(),    // mdi-access-point
-            transcribing: "\u{f04ce}".to_string(), // mdi-sync
-            stopped: "\u{f036d}".to_string(),      // mdi-microphone-off
-        },
-        "phosphor" => ResolvedIcons {
-            // Phosphor Icons (requires Phosphor font)
-            idle: "\u{e43a}".to_string(),         // ph-microphone
-            recording: "\u{e438}".to_string(),    // ph-record
-            streaming: "\u{e7ee}".to_string(),    // ph-broadcast
-            transcribing: "\u{e225}".to_string(), // ph-circle-notch (spinner)
-            stopped: "\u{e43b}".to_string(),      // ph-microphone-slash
-        },
-        "codicons" => ResolvedIcons {
-            // VS Code Codicons (requires Codicons font)
-            idle: "\u{eb51}".to_string(),         // codicon-mic
-            recording: "\u{ebfc}".to_string(),    // codicon-record
-            streaming: "\u{ebba}".to_string(),    // codicon-radio-tower
-            transcribing: "\u{eb4c}".to_string(), // codicon-sync
-            stopped: "\u{eb52}".to_string(),      // codicon-mute
-        },
-        "text" => ResolvedIcons {
-            // Plain text labels (no special fonts required)
-            idle: "[MIC]".to_string(),
-            recording: "[REC]".to_string(),
-            streaming: "[LIVE]".to_string(),
-            transcribing: "[...]".to_string(),
-            stopped: "[OFF]".to_string(),
-        },
-        "dots" => ResolvedIcons {
-            // Unicode geometric shapes (no special fonts required)
-            idle: "◯".to_string(),         // U+25EF white circle
-            recording: "⬤".to_string(),    // U+2B24 black large circle
-            streaming: "⊙".to_string(),    // U+2299 circled dot operator
-            transcribing: "◔".to_string(), // U+25D4 circle with upper right quadrant black
-            stopped: "◌".to_string(),      // U+25CC dotted circle
-        },
-        "arrows" => ResolvedIcons {
-            // Media player style (no special fonts required)
-            idle: "▶".to_string(),         // U+25B6 play
-            recording: "●".to_string(),    // U+25CF black circle
-            streaming: "⇉".to_string(),    // U+21C9 paired rightward arrows — flow
-            transcribing: "↻".to_string(), // U+21BB clockwise arrow
-            stopped: "■".to_string(),      // U+25A0 black square
-        },
-        path => load_custom_icon_theme(path).unwrap_or_else(|e| {
-            tracing::warn!(
-                "Failed to load custom icon theme '{}': {}, using emoji",
-                path,
-                e
-            );
-            load_icon_theme("emoji")
-        }),
-    }
-}
-
-/// Load a custom icon theme from a TOML file
-fn load_custom_icon_theme(path: &str) -> Result<ResolvedIcons, String> {
-    let path = PathBuf::from(path);
-    if !path.exists() {
-        return Err(format!("Theme file not found: {}", path.display()));
-    }
-
-    let contents =
-        std::fs::read_to_string(&path).map_err(|e| format!("Failed to read theme file: {}", e))?;
-
-    #[derive(Deserialize)]
-    struct ThemeFile {
-        idle: Option<String>,
-        recording: Option<String>,
-        streaming: Option<String>,
-        transcribing: Option<String>,
-        stopped: Option<String>,
-    }
-
-    let theme: ThemeFile =
-        toml::from_str(&contents).map_err(|e| format!("Invalid theme file: {}", e))?;
-
-    // Start with emoji defaults, override with file values
-    let base = load_icon_theme("emoji");
-    Ok(ResolvedIcons {
-        idle: theme.idle.unwrap_or(base.idle),
-        recording: theme.recording.unwrap_or(base.recording),
-        streaming: theme.streaming.unwrap_or(base.streaming),
-        transcribing: theme.transcribing.unwrap_or(base.transcribing),
-        stopped: theme.stopped.unwrap_or(base.stopped),
-    })
-}
-
-/// Whisper execution mode (how whisper runs)
-#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum WhisperMode {
-    /// Local transcription using whisper.cpp FFI
-    #[default]
-    Local,
-    /// Remote transcription via OpenAI-compatible API
-    Remote,
-    /// CLI transcription using whisper-cli subprocess
-    /// Fallback for systems where whisper-rs FFI doesn't work (e.g., glibc 2.42+)
-    Cli,
-}
-
-/// Language configuration supporting single language or array of allowed languages
-///
-/// Supports three modes:
-/// - Single language: `language = "en"` - use this specific language
-/// - Auto-detect: `language = "auto"` - let Whisper detect from all languages
-/// - Constrained auto-detect: `language = ["en", "fr"]` - detect from allowed set
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum LanguageConfig {
-    /// Single language code (e.g., "en", "fr", "auto")
-    Single(String),
-    /// Array of allowed language codes for constrained auto-detection
-    Multiple(Vec<String>),
-}
-
-impl Default for LanguageConfig {
-    fn default() -> Self {
-        LanguageConfig::Single("en".to_string())
-    }
-}
-
-impl LanguageConfig {
-    /// Convert to a vector of language codes
-    pub fn as_vec(&self) -> Vec<String> {
-        match self {
-            LanguageConfig::Single(s) => vec![s.clone()],
-            LanguageConfig::Multiple(v) => v.clone(),
-        }
-    }
-
-    /// Check if this is the "auto" setting (unconstrained auto-detection)
-    pub fn is_auto(&self) -> bool {
-        matches!(self, LanguageConfig::Single(s) if s == "auto")
-    }
-
-    /// Check if multiple languages are configured (constrained auto-detection)
-    pub fn is_multiple(&self) -> bool {
-        matches!(self, LanguageConfig::Multiple(v) if v.len() > 1)
-    }
-
-    /// Get the first/primary language (used for fallback or single-language mode)
-    pub fn primary(&self) -> &str {
-        match self {
-            LanguageConfig::Single(s) => s,
-            LanguageConfig::Multiple(v) => v.first().map(|s| s.as_str()).unwrap_or("en"),
-        }
-    }
-
-    /// Parse from a comma-separated string (used for CLI argument passing)
-    ///
-    /// Examples:
-    /// - "en" -> Single("en")
-    /// - "auto" -> Single("auto")
-    /// - "en,fr,de" -> Multiple(["en", "fr", "de"])
-    pub fn from_comma_separated(s: &str) -> Self {
-        let parts: Vec<String> = s.split(',').map(|p| p.trim().to_string()).collect();
-        if parts.len() == 1 {
-            LanguageConfig::Single(parts.into_iter().next().unwrap())
-        } else {
-            LanguageConfig::Multiple(parts)
-        }
-    }
-}
-
-/// Whisper speech-to-text configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct WhisperConfig {
-    /// Execution mode: "local" or "remote" (preferred field name)
-    #[serde(default)]
-    pub mode: Option<WhisperMode>,
-
-    /// DEPRECATED: Use `mode` instead. Kept for backwards compatibility.
-    #[serde(default)]
-    pub backend: Option<WhisperMode>,
-
-    /// Model name: tiny, base, small, medium, large-v3, large-v3-turbo
-    /// Can also be an absolute path to a .bin file
-    #[serde(default = "default_whisper_model")]
-    pub model: String,
-
-    /// Language configuration: single code, "auto", or array of allowed languages
-    /// Examples: "en", "auto", ["en", "fr"]
-    #[serde(default)]
-    pub language: LanguageConfig,
-
-    /// Translate to English if source language is not English
-    #[serde(default)]
-    pub translate: bool,
-
-    /// Number of threads for inference (None = auto-detect)
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-
-    /// Enable GPU memory isolation mode (default: false)
-    /// When true, transcription runs in a subprocess that exits after each
-    /// transcription, ensuring GPU memory is fully released between recordings.
-    /// This is especially useful on laptops with hybrid graphics to prevent
-    /// the GPU from staying active when not in use.
-    /// Note: This option only applies when mode = "local".
-    #[serde(default)]
-    pub gpu_isolation: bool,
-
-    /// GPU device index for Vulkan/CUDA/Metal backend selection.
-    /// On multi-GPU systems, whisper.cpp may select the integrated GPU (index 0)
-    /// instead of the discrete GPU, causing slower transcription.
-    /// Set this to the index of your preferred GPU (e.g., 1 for the second device).
-    /// Leave unset to use the default device (index 0).
-    /// You can also use the GGML_VK_VISIBLE_DEVICES env var for Vulkan filtering.
-    #[serde(default)]
-    pub gpu_device: Option<i32>,
-
-    /// Enable flash attention for GPU inference (default: false)
-    /// Reduces memory bandwidth pressure in the attention layers.
-    /// Requires a compatible GPU backend (CUDA or Vulkan).
-    #[serde(default)]
-    pub flash_attention: bool,
-
-    /// Optimize context window for short recordings (default: true)
-    /// When enabled, uses a smaller context window proportional to audio length
-    /// for clips under 22.5 seconds. This significantly speeds up transcription
-    /// on both CPU and GPU. If transcription seems unstable, set this to false.
-    #[serde(default = "default_context_window_optimization")]
-    pub context_window_optimization: bool,
-
-    // --- Eager processing settings ---
-    /// Enable eager input processing (transcribe chunks while recording continues)
-    /// When enabled, audio is split into chunks and transcribed in parallel with
-    /// continued recording. This reduces perceived latency on slower machines.
-    #[serde(default)]
-    pub eager_processing: bool,
-
-    /// Duration of each audio chunk in seconds for eager processing
-    #[serde(default = "default_eager_chunk_secs")]
-    pub eager_chunk_secs: f32,
-
-    /// Overlap between adjacent chunks in seconds for eager processing
-    /// Overlap helps catch words at chunk boundaries
-    #[serde(default = "default_eager_overlap_secs")]
-    pub eager_overlap_secs: f32,
-
-    /// Initial prompt to provide context for transcription
-    /// Use this to hint at terminology, proper nouns, or formatting conventions.
-    /// Example: "Technical discussion about Rust, TypeScript, and Kubernetes."
-    #[serde(default)]
-    pub initial_prompt: Option<String>,
-
-    // --- Multi-model settings ---
-    /// Secondary model to use when hotkey.model_modifier is held
-    /// Example: "large-v3-turbo" for difficult audio
-    #[serde(default)]
-    pub secondary_model: Option<String>,
-
-    /// List of available models that can be selected via CLI --model flag
-    /// These models can be loaded on-demand when requested
-    #[serde(default)]
-    pub available_models: Vec<String>,
-
-    /// Maximum number of models to keep loaded in memory (LRU eviction)
-    /// Default: 2 (primary model + one secondary)
-    /// Only applies when gpu_isolation = false
-    #[serde(default = "default_max_loaded_models")]
-    pub max_loaded_models: usize,
-
-    /// Seconds before unloading idle secondary models from memory
-    /// Default: 300 (5 minutes). Set to 0 to never auto-unload.
-    /// Only applies when gpu_isolation = false
-    #[serde(default = "default_cold_model_timeout")]
-    pub cold_model_timeout_secs: u64,
-
-    // --- Remote backend settings ---
-    /// Remote server endpoint URL (e.g., "http://192.168.1.100:8080")
-    /// Required when mode = "remote"
-    #[serde(default)]
-    pub remote_endpoint: Option<String>,
-
-    /// Model name to send to remote server (default: "whisper-1")
-    #[serde(default)]
-    pub remote_model: Option<String>,
-
-    /// API key for remote server (optional, can also use VOXTYPE_WHISPER_API_KEY env var)
-    #[serde(default)]
-    pub remote_api_key: Option<String>,
-
-    /// Timeout for remote requests in seconds (default: 30)
-    #[serde(default)]
-    pub remote_timeout_secs: Option<u64>,
-
-    // --- CLI backend settings ---
-    /// Path to whisper-cli binary (optional, searches PATH if not set)
-    /// Used when mode = "cli"
-    #[serde(default)]
-    pub whisper_cli_path: Option<String>,
-}
-
-impl WhisperConfig {
-    /// Get the effective execution mode, preferring `mode` over deprecated `backend`
-    pub fn effective_mode(&self) -> WhisperMode {
-        // Prefer `mode` if set
-        if let Some(mode) = self.mode {
-            return mode;
-        }
-        // Fall back to deprecated `backend` with warning
-        if let Some(backend) = self.backend {
-            tracing::warn!("DEPRECATED: [whisper] backend is deprecated, use 'mode' instead");
-            tracing::warn!(
-                "  Change 'backend = \"{}\"' to 'mode = \"{}\"' in config.toml",
-                match backend {
-                    WhisperMode::Local => "local",
-                    WhisperMode::Remote => "remote",
-                    WhisperMode::Cli => "cli",
-                },
-                match backend {
-                    WhisperMode::Local => "local",
-                    WhisperMode::Remote => "remote",
-                    WhisperMode::Cli => "cli",
-                }
-            );
-            return backend;
-        }
-        WhisperMode::default()
-    }
-}
-
-impl Default for WhisperConfig {
-    fn default() -> Self {
-        Self {
-            mode: None,    // Defaults to Local via effective_mode()
-            backend: None, // Deprecated alias
-            model: "base.en".to_string(),
-            language: LanguageConfig::default(),
-            translate: false,
-            threads: None,
-            on_demand_loading: default_on_demand_loading(),
-            gpu_isolation: false,
-            gpu_device: None,
-            flash_attention: false,
-            context_window_optimization: default_context_window_optimization(),
-            eager_processing: false,
-            eager_chunk_secs: default_eager_chunk_secs(),
-            eager_overlap_secs: default_eager_overlap_secs(),
-            initial_prompt: None,
-            secondary_model: None,
-            available_models: vec![],
-            max_loaded_models: default_max_loaded_models(),
-            cold_model_timeout_secs: default_cold_model_timeout(),
-            remote_endpoint: None,
-            remote_model: None,
-            remote_api_key: None,
-            remote_timeout_secs: None,
-            whisper_cli_path: None,
-        }
-    }
-}
-
-/// Parakeet model architecture type
-#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum ParakeetModelType {
-    /// CTC (Connectionist Temporal Classification) - faster, character-level output
-    Ctc,
-    /// TDT (Token-Duration-Transducer) - recommended, proper punctuation and word boundaries
-    #[default]
-    Tdt,
-}
-
-/// Parakeet speech-to-text configuration (ONNX-based, alternative to Whisper)
-/// Requires: cargo build --features parakeet
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct ParakeetConfig {
-    /// Path to model directory containing ONNX model files
-    /// For TDT: encoder-model.onnx, decoder_joint-model.onnx, vocab.txt
-    /// For CTC: model.onnx, tokenizer.json
-    pub model: String,
-
-    /// Model architecture type: "tdt" (default, recommended) or "ctc"
-    /// Auto-detected from model directory structure if not specified
-    #[serde(default)]
-    pub model_type: Option<ParakeetModelType>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-
-    /// Use the cache-aware streaming pipeline (parakeet-rs `ParakeetUnified`)
-    /// instead of the batch CTC/TDT models. When true, voxtype emits live
-    /// partial transcripts during recording and types the final transcript
-    /// on hotkey release. Requires a streaming-capable model directory
-    /// (TDT v3 family with `tokenizer.model`).
-    /// Default: false (batch pipeline, identical to pre-streaming behavior).
-    #[serde(default)]
-    pub streaming: bool,
-
-    /// Streaming chunk length in seconds. Smaller = lower latency, more
-    /// inference overhead. Maps to `UnifiedStreamingConfig::chunk_secs`.
-    #[serde(default = "default_streaming_chunk_secs")]
-    pub streaming_chunk_secs: f32,
-
-    /// Streaming left context in seconds. Maps to
-    /// `UnifiedStreamingConfig::left_context_secs`.
-    #[serde(default = "default_streaming_left_context_secs")]
-    pub streaming_left_context_secs: f32,
-
-    /// Streaming right context in seconds. Maps to
-    /// `UnifiedStreamingConfig::right_context_secs`.
-    #[serde(default = "default_streaming_right_context_secs")]
-    pub streaming_right_context_secs: f32,
-}
-
-fn default_streaming_chunk_secs() -> f32 {
-    0.5
-}
-
-fn default_streaming_left_context_secs() -> f32 {
-    1.5
-}
-
-fn default_streaming_right_context_secs() -> f32 {
-    0.5
-}
-
-impl Default for ParakeetConfig {
-    fn default() -> Self {
-        Self {
-            model: "parakeet-tdt-0.6b-v3".to_string(),
-            model_type: None, // Auto-detect
-            on_demand_loading: false,
-            streaming: false,
-            streaming_chunk_secs: default_streaming_chunk_secs(),
-            streaming_left_context_secs: default_streaming_left_context_secs(),
-            streaming_right_context_secs: default_streaming_right_context_secs(),
-        }
-    }
-}
-
-/// Moonshine speech-to-text configuration (ONNX-based, encoder-decoder ASR)
-/// Requires: cargo build --features moonshine
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MoonshineConfig {
-    /// Model name or path to directory containing ONNX model files
-    /// Expects: encoder_model.onnx, decoder_model_merged.onnx, tokenizer.json
-    /// Short names: "tiny" (27M params), "base" (61M params)
-    pub model: String,
-
-    /// Use quantized model variants for faster CPU inference (default: true)
-    /// Falls back to full precision if quantized files are not found
-    #[serde(default = "default_true")]
-    pub quantized: bool,
-
-    /// Number of CPU threads for ONNX Runtime inference
-    #[serde(default)]
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-}
-
-impl Default for MoonshineConfig {
-    fn default() -> Self {
-        Self {
-            model: "base".to_string(),
-            quantized: true,
-            threads: None,
-            on_demand_loading: false,
-        }
-    }
-}
-
-/// Cohere Transcribe speech-to-text configuration (ONNX-based, encoder-decoder).
-/// Requires: cargo build --features cohere
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct CohereConfig {
-    /// Model name or directory containing the Cohere ONNX files.
-    /// Expects HuggingFace Optimum layout:
-    ///   encoder_model.onnx (+ .onnx_data),
-    ///   decoder_model_merged.onnx (+ .onnx_data),
-    ///   tokenizer.json
-    /// Short names: "cohere-transcribe-q4f16" (default, ~1.5 GB),
-    ///              "cohere-transcribe-q4", "cohere-transcribe-int8",
-    ///              "cohere-transcribe-fp16"
-    pub model: String,
-
-    /// Language for transcription. Two-letter ISO 639-1 codes
-    /// (e.g. "en", "fr", "de"). Cohere supports 14 languages.
-    #[serde(default = "default_cohere_language")]
-    pub language: String,
-
-    /// Number of CPU threads for ONNX Runtime inference
-    #[serde(default)]
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-}
-
-fn default_cohere_language() -> String {
-    "en".to_string()
-}
-
-impl Default for CohereConfig {
-    fn default() -> Self {
-        Self {
-            model: "cohere-transcribe-q4f16".to_string(),
-            language: default_cohere_language(),
-            threads: None,
-            on_demand_loading: false,
-        }
-    }
-}
-
-/// Soniox cloud streaming WebSocket STT configuration
-/// Requires: cargo build --features soniox
-///
-/// Soniox is a paid cloud STT provider. API key required:
-/// either set `api_key` here or via the `SONIOX_API_KEY` env var.
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct SonioxConfig {
-    /// API key. If unset, falls back to the SONIOX_API_KEY env var.
-    #[serde(default)]
-    pub api_key: Option<String>,
-
-    /// Soniox model name. Default: "stt-rt-v4".
-    #[serde(default = "default_soniox_model")]
-    pub model: String,
-
-    /// Language hints (ISO 639-1 codes). Default: ["hu", "en"].
-    /// Empty array means auto-detect.
-    #[serde(default = "default_soniox_language_hints")]
-    pub language_hints: Vec<String>,
-
-    /// Strictly restrict recognition to the languages in `language_hints`.
-    /// When true (default), the model strongly prefers producing output
-    /// only in the hinted languages, avoiding occasional drift to a third
-    /// language in mid-stream partials. Ignored when `language_hints` is
-    /// empty. See https://soniox.com/docs/stt/concepts/language-restrictions.
-    #[serde(default = "default_true")]
-    pub language_hints_strict: bool,
-
-    /// Streaming mode. true = WebSocket session with live partials
-    /// (requires [hotkey] mode = "toggle"; PTT auto-promoted to toggle).
-    /// false = batch mode: buffer audio while held, send one-shot on release
-    /// (PTT-compatible).
-    #[serde(default = "default_true")]
-    pub streaming: bool,
-
-    /// Type partials at cursor as they arrive (streaming mode only).
-    /// false = only finalized segments are typed. Default: true.
-    #[serde(default = "default_true")]
-    pub type_partials: bool,
-
-    /// Free-form context text — mapped to `context.text` in Soniox's init
-    /// frame. Use for short domain prose ("medical consultation",
-    /// "podcast about Rust async runtime"). See
-    /// https://soniox.com/docs/stt/concepts/context.
-    #[serde(default)]
-    pub context: Option<String>,
-
-    /// Vocabulary boost terms (proper names, jargon, product names).
-    /// Mapped to `context.terms` in Soniox's init frame. Can be combined
-    /// with `terms_file`; entries are deduplicated in order.
-    #[serde(default)]
-    pub terms: Option<Vec<String>>,
-
-    /// Path to a JSON file containing a list of vocabulary boost terms
-    /// (`["term1", "term2", ...]`). Loaded once at daemon startup and
-    /// merged into `context.terms`. Useful for sharing a single
-    /// corrections list across multiple voxtype config snapshots.
-    #[serde(default)]
-    pub terms_file: Option<std::path::PathBuf>,
-
-    /// Use the Soniox async transcription API (file upload + poll) instead
-    /// of the realtime WebSocket. Higher accuracy, PTT-compatible, batch
-    /// only (no live partials). When true, overrides `streaming` and
-    /// `type_partials`. Default model becomes `stt-async-v4`.
-    /// Default: false.
-    #[serde(default)]
-    pub async_api: bool,
-
-    /// Maximum total wait time (seconds) for an async API job to complete
-    /// before giving up. Default: 120.
-    #[serde(default = "default_soniox_async_max_wait_secs")]
-    pub async_max_wait_secs: u64,
-}
-
-fn default_soniox_model() -> String {
-    "stt-rt-v4".to_string()
-}
-
-fn default_soniox_language_hints() -> Vec<String> {
-    vec!["hu".to_string(), "en".to_string()]
-}
-
-fn default_soniox_async_max_wait_secs() -> u64 {
-    120
-}
-
-impl Default for SonioxConfig {
-    fn default() -> Self {
-        Self {
-            api_key: None,
-            model: default_soniox_model(),
-            language_hints: default_soniox_language_hints(),
-            language_hints_strict: true,
-            streaming: true,
-            type_partials: true,
-            context: None,
-            terms: None,
-            terms_file: None,
-            async_api: false,
-            async_max_wait_secs: default_soniox_async_max_wait_secs(),
-        }
-    }
-}
-
-/// SenseVoice speech-to-text configuration (ONNX-based, CTC encoder-only ASR)
-/// Requires: cargo build --features sensevoice
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct SenseVoiceConfig {
-    /// Model name or path to directory containing ONNX model files
-    /// Expects: model.int8.onnx (or model.onnx), tokens.txt
-    /// Short name: "sensevoice-small" (default)
-    pub model: String,
-
-    /// Language for transcription: "auto", "zh", "en", "ja", "ko", "yue" (default: "auto")
-    #[serde(default = "default_sensevoice_language")]
-    pub language: String,
-
-    /// Enable inverse text normalization (adds punctuation) (default: true)
-    #[serde(default = "default_true")]
-    pub use_itn: bool,
-
-    /// Number of CPU threads for ONNX Runtime inference
-    #[serde(default)]
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-}
-
-fn default_sensevoice_language() -> String {
-    "auto".to_string()
-}
-
-impl Default for SenseVoiceConfig {
-    fn default() -> Self {
-        Self {
-            model: "sensevoice-small".to_string(),
-            language: "auto".to_string(),
-            use_itn: true,
-            threads: None,
-            on_demand_loading: false,
-        }
-    }
-}
-
-/// Paraformer speech-to-text configuration (FunASR ONNX-based CTC encoder)
-/// Requires: cargo build --features paraformer
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct ParaformerConfig {
-    /// Model name or path to ONNX model directory
-    /// Expects: model.onnx (or model.int8.onnx), tokens.txt
-    pub model: String,
-
-    /// Number of CPU threads for ONNX Runtime inference
-    #[serde(default)]
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-}
-
-impl Default for ParaformerConfig {
-    fn default() -> Self {
-        Self {
-            model: "paraformer-zh".to_string(),
-            threads: None,
-            on_demand_loading: false,
-        }
-    }
-}
-
-/// Dolphin speech-to-text configuration (ONNX-based CTC encoder, dictation-optimized)
-/// Requires: cargo build --features dolphin
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct DolphinConfig {
-    /// Model name or path to ONNX model directory
-    pub model: String,
-
-    /// Number of CPU threads for ONNX Runtime inference
-    #[serde(default)]
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-}
-
-impl Default for DolphinConfig {
-    fn default() -> Self {
-        Self {
-            model: "dolphin-base".to_string(),
-            threads: None,
-            on_demand_loading: false,
-        }
-    }
-}
-
-/// Omnilingual speech-to-text configuration (FunASR ONNX-based, 50+ languages)
-/// Requires: cargo build --features omnilingual
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct OmnilingualConfig {
-    /// Model name or path to ONNX model directory
-    pub model: String,
-
-    /// Number of CPU threads for ONNX Runtime inference
-    #[serde(default)]
-    pub threads: Option<usize>,
-
-    /// Load model on-demand when recording starts (true) or keep loaded (false)
-    #[serde(default = "default_on_demand_loading")]
-    pub on_demand_loading: bool,
-}
-
-impl Default for OmnilingualConfig {
-    fn default() -> Self {
-        Self {
-            model: "omnilingual-large".to_string(),
-            threads: None,
-            on_demand_loading: false,
-        }
-    }
-}
-
-/// Transcription engine selection (which ASR technology to use)
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    Deserialize,
-    Serialize,
-    PartialEq,
-    Eq,
-    Default,
-    strum::IntoStaticStr,
-    strum::Display,
-)]
-#[serde(rename_all = "lowercase")]
-#[strum(serialize_all = "lowercase")]
-pub enum TranscriptionEngine {
-    /// Use Whisper (whisper.cpp via whisper-rs)
-    #[default]
-    Whisper,
-    /// Use Parakeet (NVIDIA's FastConformer via ONNX Runtime)
-    /// Requires: cargo build --features parakeet
-    Parakeet,
-    /// Use Moonshine (encoder-decoder ASR via ONNX Runtime)
-    /// Requires: cargo build --features moonshine
-    Moonshine,
-    /// Use SenseVoice (Alibaba FunAudioLLM CTC model via ONNX Runtime)
-    /// Requires: cargo build --features sensevoice
-    SenseVoice,
-    /// Use Paraformer (FunASR CTC encoder via ONNX Runtime)
-    /// Requires: cargo build --features paraformer
-    Paraformer,
-    /// Use Dolphin (dictation-optimized CTC encoder via ONNX Runtime)
-    /// Requires: cargo build --features dolphin
-    Dolphin,
-    /// Use Omnilingual (FunASR 50+ language CTC encoder via ONNX Runtime)
-    /// Requires: cargo build --features omnilingual
-    Omnilingual,
-    /// Use Cohere Transcribe (encoder-decoder via ONNX Runtime, Whisper-style
-    /// task tokens). Top of the Open ASR Leaderboard.
-    /// Requires: cargo build --features cohere
-    Cohere,
-    /// Use Soniox (cloud streaming WebSocket STT).
-    /// Requires: cargo build --features soniox
-    Soniox,
-}
-
-impl TranscriptionEngine {
-    /// Canonical lowercase name, matching this enum's serde representation.
-    /// Backed by `strum::IntoStaticStr` so a new variant added to the enum
-    /// picks up its `serialize_all = "lowercase"` name automatically — no
-    /// match block to keep in sync. The hand-written `name()` method
-    /// remains as a stable public API so call sites read `engine.name()`
-    /// rather than the less-obvious `(*engine).into()`.
-    pub fn name(self) -> &'static str {
-        self.into()
-    }
-}
-
-/// VAD backend selection
-///
-/// Determines which voice activity detection algorithm to use.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "lowercase")]
-pub enum VadBackend {
-    /// Auto-select based on transcription engine (default)
-    /// - Whisper engine: uses Whisper VAD (requires model download)
-    /// - Parakeet engine: uses Energy VAD (no model needed)
-    #[default]
-    Auto,
-    /// Energy-based VAD using RMS amplitude analysis
-    /// Fast, no model download required, works with any engine
-    Energy,
-    /// Whisper VAD using whisper-rs built-in Silero model (GGML format)
-    /// More accurate but requires downloading ggml-silero-vad.bin
-    Whisper,
-}
-
-/// Voice Activity Detection configuration
-///
-/// VAD filters silence-only recordings before transcription to prevent
-/// Whisper hallucinations when processing silence.
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct VadConfig {
-    /// Enable Voice Activity Detection (default: false)
-    /// When enabled, recordings with no detected speech are rejected before transcription
-    #[serde(default)]
-    pub enabled: bool,
-
-    /// VAD backend to use (default: auto)
-    /// - auto: Whisper VAD for Whisper engine, Energy VAD for Parakeet
-    /// - energy: Simple RMS-based detection, no model needed
-    /// - whisper: Silero VAD via whisper-rs, requires model download
-    #[serde(default)]
-    pub backend: VadBackend,
-
-    /// Speech detection threshold (0.0-1.0, default: 0.5)
-    /// Higher values require more confident speech detection
-    #[serde(default = "default_vad_threshold")]
-    pub threshold: f32,
-
-    /// Minimum speech duration in milliseconds (default: 100)
-    /// Recordings with less speech than this are rejected
-    #[serde(default = "default_min_speech_duration_ms")]
-    pub min_speech_duration_ms: u32,
-
-    /// Path to VAD model file (optional, for Whisper VAD backend)
-    /// If not set, uses the default model location (~/.local/share/voxtype/models/)
-    #[serde(default)]
-    pub model: Option<String>,
-}
-
-fn default_vad_threshold() -> f32 {
-    0.5
-}
-
-fn default_min_speech_duration_ms() -> u32 {
-    100
-}
-
-impl Default for VadConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            backend: VadBackend::default(),
-            threshold: default_vad_threshold(),
-            min_speech_duration_ms: default_min_speech_duration_ms(),
-            model: None,
-        }
-    }
-}
-
-/// Text processing configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct TextConfig {
-    /// Enable spoken punctuation conversion (e.g., "period" → ".")
-    #[serde(default)]
-    pub spoken_punctuation: bool,
-
-    /// Custom word replacements (case-insensitive)
-    /// Example: { "vox type" = "voxtype" }
-    #[serde(default)]
-    pub replacements: HashMap<String, String>,
-
-    /// Smart auto-submit: say "submit" at the end of dictation to press Enter.
-    /// The word "submit" is stripped from the output and Enter is pressed.
-    #[serde(default)]
-    pub smart_auto_submit: bool,
-
-    /// Remove common filler words ("uh", "um", etc.) from transcribed text.
-    /// Defaults to false to preserve existing behavior. The list is
-    /// configurable via `filler_words`.
-    #[serde(default)]
-    pub filter_filler_words: bool,
-
-    /// Words removed when `filter_filler_words` is true. Matched
-    /// case-insensitively on word boundaries; surrounding punctuation and
-    /// whitespace are cleaned up after removal.
-    #[serde(default = "default_filler_words")]
-    pub filler_words: Vec<String>,
-}
-
-impl Default for TextConfig {
-    fn default() -> Self {
-        Self {
-            spoken_punctuation: false,
-            replacements: HashMap::new(),
-            smart_auto_submit: false,
-            filter_filler_words: true,
-            filler_words: default_filler_words(),
-        }
-    }
-}
-
-/// Default filler-word list. Conservative: single-syllable disfluencies only.
-/// Multi-word phrases like "you know" or "sort of" are too aggressive for a
-/// default and can be added via the `filler_words` config.
-fn default_filler_words() -> Vec<String> {
-    vec![
-        "uh".to_string(),
-        "um".to_string(),
-        "er".to_string(),
-        "ah".to_string(),
-        "eh".to_string(),
-        "hmm".to_string(),
-        "hm".to_string(),
-        "mm".to_string(),
-        "mhm".to_string(),
-    ]
-}
-
-/// Meeting transcription configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MeetingConfig {
-    /// Enable meeting mode
-    #[serde(default)]
-    pub enabled: bool,
-
-    /// Duration of each audio chunk in seconds
-    #[serde(default = "default_chunk_duration")]
-    pub chunk_duration_secs: u32,
-
-    /// Storage path for meetings ("auto" for default location)
-    /// Default: ~/.local/share/voxtype/meetings/
-    #[serde(default = "default_storage_path")]
-    pub storage_path: String,
-
-    /// Retain raw audio files after transcription
-    #[serde(default)]
-    pub retain_audio: bool,
-
-    /// Maximum meeting duration in minutes (0 = unlimited)
-    #[serde(default = "default_max_duration")]
-    pub max_duration_mins: u32,
-
-    /// Meeting audio configuration
-    #[serde(default)]
-    pub audio: MeetingAudioConfig,
-
-    /// Diarization configuration
-    #[serde(default)]
-    pub diarization: MeetingDiarizationConfig,
-
-    /// Summarization configuration
-    #[serde(default)]
-    pub summary: MeetingSummaryConfig,
-}
-
-/// Meeting audio configuration for dual capture
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MeetingAudioConfig {
-    /// Microphone device (uses main audio.device if not specified)
-    #[serde(default = "default_mic_device")]
-    pub mic_device: String,
-
-    /// Loopback device for capturing remote participants
-    /// Options: "auto" (detect), "disabled", or specific device name
-    #[serde(default = "default_loopback")]
-    pub loopback_device: String,
-
-    /// Echo cancellation mode for removing speaker bleed-through from mic
-    /// Options: "auto" (GTCRN neural enhancement + transcript dedup), "disabled"
-    /// The GTCRN model (~523KB) is auto-downloaded on first meeting start.
-    /// For system-level echo cancellation, configure PipeWire's echo-cancel module
-    /// and set this to "disabled".
-    #[serde(default = "default_echo_cancel")]
-    pub echo_cancel: String,
-
-    /// RMS threshold for meeting chunk voice activity detection.
-    /// Lower values are more permissive; 0.0 disables the pre-transcription gate.
-    #[serde(default = "default_meeting_vad_threshold")]
-    pub vad_threshold: f32,
-}
-
-fn default_mic_device() -> String {
-    "default".to_string()
-}
-
-fn default_loopback() -> String {
-    "auto".to_string()
-}
-
-fn default_echo_cancel() -> String {
-    "auto".to_string()
-}
-
-fn default_meeting_vad_threshold() -> f32 {
-    0.01
-}
-
-impl Default for MeetingAudioConfig {
-    fn default() -> Self {
-        Self {
-            mic_device: default_mic_device(),
-            loopback_device: default_loopback(),
-            echo_cancel: default_echo_cancel(),
-            vad_threshold: default_meeting_vad_threshold(),
-        }
-    }
-}
-
-/// Meeting diarization configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MeetingDiarizationConfig {
-    /// Enable speaker diarization
-    #[serde(default = "default_true")]
-    pub enabled: bool,
-
-    /// Diarization backend: "simple", "ml", or "remote"
-    #[serde(default = "default_diarization_backend")]
-    pub backend: String,
-
-    /// Maximum number of speakers to detect
-    #[serde(default = "default_max_speakers")]
-    pub max_speakers: u32,
-
-    /// Path to ONNX model for ML backend (uses default if not set)
-    #[serde(default)]
-    pub model_path: Option<String>,
-
-    /// Minimum segment duration in milliseconds for ML embedding extraction
-    #[serde(default = "default_min_segment_ms")]
-    pub min_segment_ms: u64,
-
-    // The four fields below apply only to backend = "ml"; the "simple" and
-    // "remote" backends ignore them.
-    /// Cosine similarity threshold for the ML backend (0.20-0.30 typical for ECAPA on 4s windows)
-    #[serde(default = "default_similarity_threshold")]
-    pub similarity_threshold: f32,
-
-    /// VAD sub-window length in seconds for ECAPA feeding
-    #[serde(default = "default_vad_window_secs")]
-    pub vad_window_secs: f32,
-
-    /// VAD sub-window hop in seconds
-    #[serde(default = "default_vad_hop_secs")]
-    pub vad_hop_secs: f32,
-
-    /// RMS floor for treating a sub-window as silence
-    #[serde(default = "default_vad_rms_floor")]
-    pub vad_rms_floor: f32,
-}
-
-fn default_diarization_backend() -> String {
-    "simple".to_string()
-}
-
-fn default_max_speakers() -> u32 {
-    10
-}
-
-fn default_min_segment_ms() -> u64 {
-    500
-}
-
-// Empirically tuned against multi-speaker test clips (4-person roundtable,
-// 3-person panel, 1h talk). The previous 0.75 anchor was far too strict for
-// 4s ECAPA windows and produced overwhelmingly Unknown labels in practice.
-fn default_similarity_threshold() -> f32 {
-    0.25
-}
-
-fn default_vad_window_secs() -> f32 {
-    4.0
-}
-
-fn default_vad_hop_secs() -> f32 {
-    2.0
-}
-
-fn default_vad_rms_floor() -> f32 {
-    0.005
-}
-
-fn default_chunk_duration() -> u32 {
-    30
-}
-
-fn default_storage_path() -> String {
-    "auto".to_string()
-}
-
-fn default_max_duration() -> u32 {
-    180
-}
-
-impl Default for MeetingDiarizationConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            backend: default_diarization_backend(),
-            max_speakers: default_max_speakers(),
-            model_path: None,
-            min_segment_ms: default_min_segment_ms(),
-            similarity_threshold: default_similarity_threshold(),
-            vad_window_secs: default_vad_window_secs(),
-            vad_hop_secs: default_vad_hop_secs(),
-            vad_rms_floor: default_vad_rms_floor(),
-        }
-    }
-}
-
-/// Meeting summary configuration (Phase 5)
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MeetingSummaryConfig {
-    /// Summarization backend: "local", "remote", or "disabled"
-    #[serde(default = "default_summary_backend")]
-    pub backend: String,
-
-    /// Ollama URL for local backend
-    #[serde(default = "default_ollama_url")]
-    pub ollama_url: String,
-
-    /// Ollama model name
-    #[serde(default = "default_ollama_model")]
-    pub ollama_model: String,
-
-    /// Remote API endpoint for remote backend
-    #[serde(default)]
-    pub remote_endpoint: Option<String>,
-
-    /// Remote API key
-    #[serde(default)]
-    pub remote_api_key: Option<String>,
-
-    /// Request timeout in seconds
-    #[serde(default = "default_summary_timeout")]
-    pub timeout_secs: u64,
-}
-
-fn default_summary_backend() -> String {
-    "disabled".to_string()
-}
-
-fn default_ollama_url() -> String {
-    "http://localhost:11434".to_string()
-}
-
-fn default_ollama_model() -> String {
-    "llama3.2".to_string()
-}
-
-fn default_summary_timeout() -> u64 {
-    120
-}
-
-impl Default for MeetingSummaryConfig {
-    fn default() -> Self {
-        Self {
-            backend: default_summary_backend(),
-            ollama_url: default_ollama_url(),
-            ollama_model: default_ollama_model(),
-            remote_endpoint: None,
-            remote_api_key: None,
-            timeout_secs: default_summary_timeout(),
-        }
-    }
-}
-
-impl Default for MeetingConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            chunk_duration_secs: default_chunk_duration(),
-            storage_path: default_storage_path(),
-            retain_audio: false,
-            max_duration_mins: default_max_duration(),
-            audio: MeetingAudioConfig::default(),
-            diarization: MeetingDiarizationConfig::default(),
-            summary: MeetingSummaryConfig::default(),
-        }
-    }
-}
-
-/// Notification configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct NotificationConfig {
-    /// Notify when recording starts (hotkey pressed)
-    #[serde(default)]
-    pub on_recording_start: bool,
-
-    /// Notify when recording stops (hotkey released, transcription starting)
-    #[serde(default)]
-    pub on_recording_stop: bool,
-
-    /// Notify with transcribed text after transcription completes
-    #[serde(default = "default_true")]
-    pub on_transcription: bool,
-
-    /// Show engine icon in notification title (🦜 for Parakeet, 🗣️ for Whisper)
-    #[serde(default)]
-    pub show_engine_icon: bool,
-
-    /// Notification urgency level: "low", "normal", or "critical".
-    /// On GNOME, "low" notifications go straight to the drawer without a popup banner.
-    #[serde(default = "default_notification_urgency")]
-    pub urgency: String,
-}
-
-fn default_notification_urgency() -> String {
-    "normal".to_string()
-}
-
-impl Default for NotificationConfig {
-    fn default() -> Self {
-        Self {
-            on_recording_start: false,
-            on_recording_stop: false,
-            on_transcription: true,
-            show_engine_icon: false,
-            urgency: default_notification_urgency(),
-        }
-    }
-}
-
-/// Post-processing command configuration
-///
-/// Pipes transcribed text through an external command for cleanup/formatting.
-/// Commonly used with local LLMs (Ollama, llama.cpp) or text processing tools.
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct PostProcessConfig {
-    /// Shell command to execute
-    /// Receives transcribed text on stdin, outputs processed text on stdout
-    pub command: String,
-
-    /// Timeout in milliseconds (default: 30000 = 30 seconds)
-    #[serde(default = "default_post_process_timeout")]
-    pub timeout_ms: u64,
-
-    /// Whether to trim leading/trailing whitespace from command output (default: true)
-    /// Set to false when the command intentionally produces significant whitespace,
-    /// e.g. a trailing space after sentence-ending punctuation for dictation flow.
-    #[serde(default = "default_true")]
-    pub trim: bool,
-
-    /// Whether to fall back to original text when command output is empty (default: true)
-    /// Set to false when the command intentionally produces empty output,
-    /// e.g. filtering out unwanted transcriptions like [BLANK_AUDIO].
-    #[serde(default = "default_true")]
-    pub fallback_on_empty: bool,
-}
-
-/// Named profile for context-specific settings
-///
-/// Profiles allow different post-processing commands (and other settings)
-/// for different contexts like Slack, code editors, email, etc.
-///
-/// # Example Configuration
-///
-/// ```toml
-/// [profiles.slack]
-/// post_process_command = "cleanup-for-slack.sh"
-///
-/// [profiles.code]
-/// post_process_command = "cleanup-for-code.sh"
-/// ```
-///
-/// Use with: `voxtype record start --profile slack`
-#[derive(Debug, Clone, Default, Deserialize, Serialize)]
-pub struct Profile {
-    /// Post-processing command for this profile
-    /// Overrides [output.post_process.command] when the profile is active
-    #[serde(default)]
-    pub post_process_command: Option<String>,
-
-    /// Timeout for post-processing in milliseconds (default: 30000)
-    #[serde(default)]
-    pub post_process_timeout_ms: Option<u64>,
-
-    /// Output mode override for this profile
-    #[serde(default)]
-    pub output_mode: Option<OutputMode>,
-}
-
-fn default_post_process_timeout() -> u64 {
-    30000 // 30 seconds - generous for LLM processing
-}
-
-fn default_restore_clipboard_delay() -> u32 {
-    200 // 200ms - delay for paste to complete before restoring clipboard
-}
-
-/// Text output configuration
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct OutputConfig {
-    /// Primary output mode
-    #[serde(default)]
-    pub mode: OutputMode,
-
-    /// Fall back to clipboard if typing fails
-    #[serde(default = "default_true")]
-    pub fallback_to_clipboard: bool,
-
-    /// Custom driver order for type mode (overrides default: wtype -> dotool -> ydotool -> clipboard)
-    /// Specify which drivers to try and in what order.
-    /// Example: ["ydotool", "wtype"] to prefer ydotool over wtype
-    #[serde(default)]
-    pub driver_order: Option<Vec<OutputDriver>>,
-
-    /// Notification settings
-    #[serde(default)]
-    pub notification: NotificationConfig,
-
-    /// Delay between typed characters (ms), 0 for fastest
-    #[serde(default)]
-    pub type_delay_ms: u32,
-
-    /// Delay before typing starts (ms), allows virtual keyboard to initialize
-    /// Helps prevent first character from being dropped on some compositors
-    #[serde(default)]
-    pub pre_type_delay_ms: u32,
-
-    /// DEPRECATED: Use pre_type_delay_ms instead. Kept for backwards compatibility.
-    #[serde(default)]
-    pub wtype_delay_ms: u32,
-
-    /// Automatically submit (send Enter key) after outputting transcribed text
-    /// Useful for chat applications, command lines, or forms where you want
-    /// to auto-submit after dictation
-    #[serde(default)]
-    pub auto_submit: bool,
-
-    /// Text to append after each transcription (e.g., " " for a space)
-    /// Appended after the transcription but before auto_submit
-    /// Useful for separating sentences when dictating paragraphs incrementally
-    #[serde(default)]
-    pub append_text: Option<String>,
-
-    /// Convert newlines to Shift+Enter instead of regular Enter
-    /// Useful for applications where Enter submits (e.g., Cursor IDE, Slack, Discord)
-    #[serde(default)]
-    pub shift_enter_newlines: bool,
-
-    /// Prefix wtype output with a Shift key press/release
-    /// Workaround for apps (e.g., Discord) that drop the first CJK character
-    #[serde(default)]
-    pub wtype_shift_prefix: bool,
-
-    /// Command to run when recording starts (e.g., switch to compositor submap)
-    /// Useful for entering a mode where cancel keybindings are effective
-    #[serde(default)]
-    pub pre_recording_command: Option<String>,
-
-    /// Command to run before typing output (e.g., compositor submap switch)
-    /// Useful for blocking modifier keys at the compositor level
-    #[serde(default)]
-    pub pre_output_command: Option<String>,
-
-    /// Command to run after typing output (e.g., reset compositor submap)
-    /// Runs even if typing fails, to ensure cleanup
-    #[serde(default)]
-    pub post_output_command: Option<String>,
-
-    /// Optional post-processing command configuration
-    /// Pipes transcribed text through an external command before output
-    #[serde(default)]
-    pub post_process: Option<PostProcessConfig>,
-
-    /// Keystroke to simulate for paste mode (e.g., "ctrl+v", "shift+insert", "ctrl+shift+v")
-    /// Defaults to "ctrl+v" if not specified
-    #[serde(default)]
-    pub paste_keys: Option<String>,
-
-    /// Keyboard layout for dotool (e.g., "de" for German, "fr" for French)
-    /// Required for non-US keyboard layouts when using dotool
-    #[serde(default)]
-    pub dotool_xkb_layout: Option<String>,
-
-    /// Keyboard layout variant for dotool (e.g., "nodeadkeys")
-    #[serde(default)]
-    pub dotool_xkb_variant: Option<String>,
-
-    /// Keyboard layout for eitype (e.g., "de" for German, "ru" for Russian).
-    /// Passed to eitype as `-l <layout>`. Overrides the system XKB layout
-    /// while eitype is typing, then restores it when eitype exits.
-    /// Required when the transcribed language does not match the active
-    /// system layout (see issue #180).
-    #[serde(default)]
-    pub eitype_xkb_layout: Option<String>,
-
-    /// Keyboard layout variant for eitype (e.g., "dvorak", "colemak").
-    /// Passed to eitype as `--variant <variant>`.
-    #[serde(default)]
-    pub eitype_xkb_variant: Option<String>,
-
-    /// Mapping from detected language code (two-letter ISO 639-1) to XKB
-    /// keyboard layout. When voxtype's transcriber reports a language for the
-    /// current transcription and no explicit `eitype_xkb_layout` /
-    /// `dotool_xkb_layout` is set, the layout is looked up here.
-    ///
-    /// Built-in defaults cover the common cases (en→us, ru→ru, de→de, ...);
-    /// see [`default_language_to_layout`]. Users can override or extend the
-    /// map in config to handle layouts that differ from the language code
-    /// (e.g. `pt = "br"` for Brazilian Portuguese).
-    ///
-    /// Set to an empty map (or remove all entries) to disable automatic
-    /// layout selection from the detected language.
-    #[serde(default = "default_language_to_layout")]
-    pub language_to_layout: std::collections::HashMap<String, String>,
-
-    /// Mapping from detected language code (two-letter ISO 639-1) to XKB
-    /// keyboard layout variant. Applied per transcription, after
-    /// `language_to_layout`, when no explicit `eitype_xkb_variant` /
-    /// `dotool_xkb_variant` is set.
-    ///
-    /// This is intentionally empty by default. Variants are user-specific
-    /// layout choices (for example, Russian phonetic vs standard).
-    #[serde(default)]
-    pub language_to_variant: std::collections::HashMap<String, String>,
-
-    /// File path for file output mode (required when mode = "file")
-    /// Also used as default path for --output-file CLI flag
-    #[serde(default)]
-    pub file_path: Option<PathBuf>,
-
-    /// File write mode: "overwrite" (default) or "append"
-    /// Applies to both config-based file output and --output-file CLI flag
-    #[serde(default)]
-    pub file_mode: FileMode,
-
-    /// Restore original clipboard content after paste mode completes
-    /// Saves clipboard before transcription, restores it after paste keystroke
-    #[serde(default)]
-    pub restore_clipboard: bool,
-
-    /// Delay after paste before restoring clipboard content (milliseconds)
-    /// Allows time for the paste operation to complete
-    #[serde(default = "default_restore_clipboard_delay")]
-    pub restore_clipboard_delay_ms: u32,
-
-    /// Wait for modifier keys (Ctrl/Alt/Shift/Super) to be released before
-    /// typing transcribed text. Prevents the typed letters from combining
-    /// with held modifiers and triggering compositor or application
-    /// keybindings (e.g. Super+X, Ctrl+W).
-    ///
-    /// Requires `/dev/input` access (typically `input` group membership).
-    /// Silently disabled when access is unavailable; output proceeds as
-    /// before in that case.
-    #[serde(default = "default_true")]
-    pub wait_for_modifier_release: bool,
-
-    /// Maximum time (milliseconds) to wait for modifier keys to be released
-    /// before falling back to clipboard output. Prevents a stuck modifier from
-    /// indefinitely blocking transcription delivery.
-    #[serde(default = "default_modifier_release_timeout_ms")]
-    pub modifier_release_timeout_ms: u64,
-}
-
-impl Default for OutputConfig {
-    fn default() -> Self {
-        Self {
-            mode: OutputMode::default(),
-            fallback_to_clipboard: true,
-            driver_order: None,
-            notification: NotificationConfig::default(),
-            type_delay_ms: 0,
-            pre_type_delay_ms: 0,
-            wtype_delay_ms: 0,
-            auto_submit: false,
-            append_text: None,
-            shift_enter_newlines: false,
-            wtype_shift_prefix: false,
-            pre_recording_command: None,
-            pre_output_command: None,
-            post_output_command: None,
-            post_process: None,
-            paste_keys: None,
-            dotool_xkb_layout: None,
-            dotool_xkb_variant: None,
-            eitype_xkb_layout: None,
-            eitype_xkb_variant: None,
-            language_to_layout: default_language_to_layout(),
-            language_to_variant: HashMap::new(),
-            file_path: None,
-            file_mode: FileMode::default(),
-            restore_clipboard: false,
-            restore_clipboard_delay_ms: default_restore_clipboard_delay(),
-            wait_for_modifier_release: true,
-            modifier_release_timeout_ms: default_modifier_release_timeout_ms(),
-        }
-    }
-}
-
-fn default_modifier_release_timeout_ms() -> u64 {
-    750
-}
-
-/// Result of applying a per-language XKB layout/variant hint to output config.
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct AppliedLanguageXkbHint {
-    /// Layout found in `language_to_layout`, if any.
-    pub layout: Option<String>,
-    /// Variant found in `language_to_variant`, if any.
-    pub variant: Option<String>,
-    /// Whether the layout was applied to eitype for this transcription.
-    pub eitype_layout_applied: bool,
-    /// Whether the layout was applied to dotool for this transcription.
-    pub dotool_layout_applied: bool,
-    /// Whether the variant was applied to eitype for this transcription.
-    pub eitype_variant_applied: bool,
-    /// Whether the variant was applied to dotool for this transcription.
-    pub dotool_variant_applied: bool,
-}
-
-impl AppliedLanguageXkbHint {
-    pub fn is_empty(&self) -> bool {
-        self.layout.is_none() && self.variant.is_none()
-    }
-}
-
-/// Built-in mapping from two-letter ISO 639-1 language codes to XKB layout
-/// codes. Used when the transcriber reports a detected language and the user
-/// has not set an explicit `eitype_xkb_layout` / `dotool_xkb_layout`.
-///
-/// Covers the most common cases where layout code matches language code
-/// (en→us is the notable exception). Users can extend or override this map
-/// in config under `[output] language_to_layout`. To disable automatic
-/// layout selection entirely, set `language_to_layout = {}` in config.
-pub fn default_language_to_layout() -> std::collections::HashMap<String, String> {
-    let mut m = std::collections::HashMap::new();
-    // English uses "us" by convention, not "en".
-    m.insert("en".to_string(), "us".to_string());
-    // Other common languages where layout name matches ISO 639-1.
-    for code in [
-        "ru", "de", "fr", "es", "it", "pl", "uk", "cs", "sk", "sv", "no", "fi", "da", "nl", "pt",
-        "tr", "gr", "hu", "ro", "bg", "hr", "sr", "sl", "lt", "lv", "et", "is", "ca", "eu",
-    ] {
-        m.insert(code.to_string(), code.to_string());
-    }
-    // Greek uses "gr" not "el".
-    m.insert("el".to_string(), "gr".to_string());
-    // Japanese, Korean, Chinese typically need IMEs rather than XKB layouts,
-    // but voxtype passes the hint through so users can map them as they wish.
-    m.insert("ja".to_string(), "jp".to_string());
-    m.insert("ko".to_string(), "kr".to_string());
-    m
-}
-
-impl OutputConfig {
-    /// Apply per-language XKB layout/variant hints to eitype and dotool.
-    ///
-    /// Explicit driver-specific settings win independently per field:
-    /// `dotool_xkb_layout` prevents only the automatic dotool layout, while
-    /// `dotool_xkb_variant` prevents only the automatic dotool variant.
-    pub fn apply_language_xkb_hint(&mut self, lang: &str) -> AppliedLanguageXkbHint {
-        let layout = self.language_to_layout.get(lang).cloned();
-        let variant = self.language_to_variant.get(lang).cloned();
-        let mut applied = AppliedLanguageXkbHint {
-            layout,
-            variant,
-            ..AppliedLanguageXkbHint::default()
-        };
-
-        if let Some(ref layout) = applied.layout {
-            if self.eitype_xkb_layout.is_none() {
-                self.eitype_xkb_layout = Some(layout.clone());
-                applied.eitype_layout_applied = true;
-            }
-            if self.dotool_xkb_layout.is_none() {
-                self.dotool_xkb_layout = Some(layout.clone());
-                applied.dotool_layout_applied = true;
-            }
-        }
-
-        if let Some(ref variant) = applied.variant {
-            if self.eitype_xkb_variant.is_none() {
-                self.eitype_xkb_variant = Some(variant.clone());
-                applied.eitype_variant_applied = true;
-            }
-            if self.dotool_xkb_variant.is_none() {
-                self.dotool_xkb_variant = Some(variant.clone());
-                applied.dotool_variant_applied = true;
-            }
-        }
-
-        applied
-    }
-
-    /// Get the effective pre-type delay, handling deprecated wtype_delay_ms
-    pub fn effective_pre_type_delay_ms(&self) -> u32 {
-        if self.wtype_delay_ms > 0 {
-            if self.pre_type_delay_ms > 0 {
-                // Both set - prefer new option, warn about deprecated
-                tracing::warn!(
-                    "Both pre_type_delay_ms and wtype_delay_ms are set. \
-                     Using pre_type_delay_ms={}. wtype_delay_ms is deprecated.",
-                    self.pre_type_delay_ms
-                );
-                self.pre_type_delay_ms
-            } else {
-                // Only deprecated option set - use it with warning
-                tracing::warn!("wtype_delay_ms is deprecated, use pre_type_delay_ms instead");
-                self.wtype_delay_ms
-            }
-        } else {
-            self.pre_type_delay_ms
-        }
-    }
-}
-
-/// Output mode selection
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum OutputMode {
-    /// Simulate keyboard input (requires ydotool)
-    #[default]
-    Type,
-    /// Copy to clipboard (wl-copy on Wayland, xclip on X11)
-    Clipboard,
-    /// Copy to clipboard then paste with Ctrl+V (requires wl-copy and ydotool)
-    Paste,
-    /// Write transcription to a file
-    File,
-}
-
-/// Output driver for typing text
-/// Used to specify preferred drivers in the fallback chain
-#[derive(Debug, Clone, Copy, Deserialize, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum OutputDriver {
-    /// wtype - Wayland-native via virtual-keyboard protocol, best Unicode/CJK support
-    Wtype,
-    /// eitype - Wayland via libei/EI protocol, works on GNOME/KDE
-    Eitype,
-    /// dotool - Works on X11/Wayland/TTY, supports keyboard layouts
-    Dotool,
-    /// ydotool - Works on X11/Wayland/TTY, requires daemon
-    Ydotool,
-    /// Clipboard via wl-copy (Wayland)
-    Clipboard,
-    /// Clipboard via xclip (X11)
-    Xclip,
-}
-
-impl std::fmt::Display for OutputDriver {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            OutputDriver::Wtype => write!(f, "wtype"),
-            OutputDriver::Eitype => write!(f, "eitype"),
-            OutputDriver::Dotool => write!(f, "dotool"),
-            OutputDriver::Ydotool => write!(f, "ydotool"),
-            OutputDriver::Clipboard => write!(f, "clipboard"),
-            OutputDriver::Xclip => write!(f, "xclip"),
-        }
-    }
-}
-
-impl std::str::FromStr for OutputDriver {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "wtype" => Ok(OutputDriver::Wtype),
-            "eitype" => Ok(OutputDriver::Eitype),
-            "dotool" => Ok(OutputDriver::Dotool),
-            "ydotool" => Ok(OutputDriver::Ydotool),
-            "clipboard" => Ok(OutputDriver::Clipboard),
-            "xclip" => Ok(OutputDriver::Xclip),
-            _ => Err(format!(
-                "Unknown driver '{}'. Valid options: wtype, eitype, dotool, ydotool, clipboard, xclip",
-                s
-            )),
-        }
-    }
-}
-
-/// File write mode when using file output
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum FileMode {
-    /// Overwrite the file on each transcription (default)
-    #[default]
-    Overwrite,
-    /// Append to the file on each transcription
-    Append,
-}
-
-fn default_true() -> bool {
-    true
 }
 
 impl Default for Config {
